Create Server
Create a simple Node server using the HTTP module. The server only listens for POST requests to the `/test` route. The route expects JSON that has a key called `string_to_cut` and returns JSON with a key called `return_string`. The API takes every 3rd of a string and returns a new string made up of those 3rd letters.